Monospace Address Bar
The use of sans-serif fonts for address bars has been a long-standing issue for me. It seems completely unsafe to use a font that doesn't distinguish between letters such as lowercase l and capital I in such a character-crucial situation. Serif fonts are slightly more acceptable, but still have problems with whitespace. This used to be less of an issue, but has grown since spaces in file paths have become more and more acceptable. I also just like monospace more.